fix: tagPage link hrefs are correctly colored
Previously, the classes of <a> links in [...tagPage].astro, such as
links to posts or prev/next links, did not follow the styles dictated in
base.css. Their text was colored as normal white text, rather than blue
normally and violet-accented upon hover.
This commit sets their classes directly, as a workaround.