feat!(dx): add CONTRIBUTING.md
* Use conventional-commits

This commit is marked as BREAKING because it deprecates the prior blog
content addition commit message in favor of a new commit type, `post`.

Before: feat(blog): content-title change
After: post(content-title): change

This doesn't break the conventional commit spec because the addition of
post content isn't exactly an addition of a feature in the codebase.